requirements.txt
yfinance
datetime
dateutil.relativedelta
plotly.express
streamlit
streamlit_plotly_events
numpy
pandas